Implement CSS parser to separate parsing logic for cleaner and reusable code
